requirements.txt
openai-whisper
Telethon
